+++ Frontend/allygator_frontend/package.json
@@ -13,10 +13,12 @@
     "react-dom": "17.0.2"
   },
   "devDependencies": {
+    "@types/react": "^17.0.38",
     "autoprefixer": "^10.4.2",
     "eslint": "8.7.0",
     "eslint-config-next": "12.0.8",
     "postcss": "^8.4.5",
-    "tailwindcss": "^3.0.15"
+    "tailwindcss": "^3.0.15",
+    "typescript": "^4.5.5"
   }
 }
