+++ Frontend/allygator_frontend/pages/index.tsx
@@ -1,0 +1,18 @@
+
+  import Head from 'next/head'
+  import Image from 'next/image'
+
+  export default function Home() {
+  return (
+    <>
+  <div className="text-3xl font-bold underline">
+    <title>AllyGators</title>
+    <meta name="description" content="Generated by create next app" />
+    <link rel="icon" href="/favicon.ico" />
+    <h2>
+      Welcome to AllyGators ...
+    </h2>
+  </div>
+  </>
+  )
+  }
